move Instance tasks to tasks module
